UScounties {spam}	R Documentation

Adjacency structure of the counties in the contiguous United States

Description

First and second order adjacency structure of the counties in the contiguous United States. We consider that two counties are neighbors if they share at least one edge of their polygon description in maps.

Format

Two matrices of class spam

UScounties.storder

Contains a one in the i and j element if county i is a neighbor of county j.

UScounties.ndorder

Contains a one in the i and j element if counties i and j are a neighbors of county k and counties i and j are not neighbors.

See Also

map from maps.

Examples

# number of counties:
n  <- nrow( UScounties.storder)

## Not run: 
# make a precision matrix
Q <- diag.spam( n) + .2 * UScounties.storder + .1 * UScounties.ndorder
display( as.spam( chol( Q)))

## End(Not run)

[Package spam version 2.5-0 Index]