USprecip {spam}	R Documentation

Monthly Total Precipitation (mm) for April 1948 in the Contiguous United States

Description

This is a useful spatial data set of moderate to large size consisting of 11918 locations. See https://www.image.ucar.edu/GSP/Data/US.monthly.met/ for the source of these data.

Format

This data set is an array containing the following columns:

lon,lat

Longitude-latitude position of monitoring stations.

raw

Monthly total precipitation in millimeters for April 1948.

anomaly

Preipitation anomaly for April 1948.

infill

Indicator, which station values were observed (5906 out of the 11918) compared to which were estimated.

Source

https://www.image.ucar.edu/GSP/Data/US.monthly.met/

References

Johns, C., Nychka, D., Kittel, T., and Daly, C. (2003) Infilling sparse records of spatial fields. Journal of the American Statistical Association, 98, 796–806.

See Also

RMprecip

Examples

# plot
## Not run: 
library(fields)

data(USprecip)
par(mfcol=c(2,1))
quilt.plot(USprecip[,1:2],USprecip[,3])
US( add=TRUE, col=2, lty=2)
quilt.plot(USprecip[,1:2],USprecip[,4])
US( add=TRUE, col=2, lty=2)

## End(Not run)

[Package spam version 2.11-3 Index]