adjacency {spam}	R Documentation

Administrative districts of Germany

Description

Constructing the adjacency graph of the administrative districts of Germany

Usage

adjacency.landkreis( loc)

Arguments

loc
location of the graph structure, can be an URL.

Details

The function is included as an example on how to construct adjacency matrices form a (common) adjacency structure. For the particular example, note that the nodes are not numbered consecutively and that they start from zero.

Value

a sparse matrix in spam format.

Author(s)

Reinhard Furrer

References

The adjacency data has been provided by Havard Rue and is also available in INLA.

See Also

germany.plot super-seeding map.landkreis for plotting.
Oral.

Examples

## Not run: 
loc <- system.file("demodata/germany.adjacency", package="spam")
display( adjacency.landkreis( loc))

## End(Not run)

[Package spam version 2.5-0 Index]