circulant {spam}	R Documentation

Create Circulant Matrices

Description

Creates a circulant matrix in spam format.

Usage

circulant.spam(x, n = NULL, eps = getOption("spam.eps"))

Arguments

x
the first row to form the circulant matrix or a list containing the indices and the nonzero values.

n
if x is a list, the dimension of the matrix.

eps
A tolerance parameter: elements of x such that abs(x) <= eps set to zero. Defaults to eps = getOption("spam.eps")

Value

The circulant matrix in spam format.

Author(s)

Reinhard Furrer

See Also

circulant from package magic, toeplitz.spam

Examples

circulant.spam(c(1,.25,0,0,0))

[Package spam version 2.3-0.1 Index]