crossprod {spam}	R Documentation

Spam Matrix Crossproduct

Description

Given matrices x and y as arguments, return a matrix cross-product. This is formally equivalent to (but usually slightly faster than) the call t(x) %*% y (crossprod.spam) or x %*% t(y) (tcrossprod.spam).

Usage

     crossprod.spam(x, y = NULL)
     
     tcrossprod.spam(x, y = NULL)

Arguments

x, y
matrices: y = NULL is taken to be the same matrix as x. Vectors are promoted to single-column or single-row matrices, depending on the context.

Value

A double matrix

Note

When x or y are not matrices, they are treated as column or row matrices.

Author(s)

Reinhard Furrer

Examples

crossprod.spam(diag.spam(2),1:2)

[Package spam version 2.3-0.1 Index]