dim {spam}	R Documentation

Dimensions of an Object

Description

Retrieve or set the dimension of an spam object.

Usage

# dim(x)
# dim(x) <- value

Arguments

x
a spam matrix

value
A numeric two-vector, which is coerced to integer (by truncation).

Details

In older version of spam, the behavior of the replacement method was different and is now implemented in pad.spam.

Value

dim retrieves the dimension slot of the object. It is a vector of mode integer.

The replacemnt method changes the dimension of the object by rearranging.

Author(s)

Reinhard Furrer

See Also

pad.spam.

Examples

x <- diag(4)
dim(x)<-c(2,8)
x

s <- diag.spam(4)
dim(s) <- c(2,8)  # result is different than x

s <- diag.spam(4)
pad(s) <- c(7,3)  # any positive value can be used

[Package spam version 2.3-0.1 Index]