foreign {spam}	R Documentation

Transformation to other sparse formats

Description

Transform between the spam sparse format to the matrix.csr format of SparseM and dgRMatrix format of Matrix

Usage

as.spam.matrix.csr(x)
as.dgRMatrix.spam(x)
as.dgCMatrix.spam(x)
as.spam.dgRMatrix(x)
as.spam.dgCMatrix(x)

Arguments

x
sparse matrix of class spam, matrix.csr, dgRMatrix or dgCMatrix.

Details

We do not provide any S4 methods and because of the existing mechanism a standard S3 does not work.

The functions are based on require.

Notice that as.matrix.csr.spam should read as as."matrix.csr".spam.

Value

According to the call, a sparse matrix of class spam, matrix.csr, dgRMatrix or dgCMatrix.

Author(s)

Reinhard Furrer

See Also

triplet, Matrix or matrix.csr

Examples

## Not run: 
S <- diag.spam(4)
R <- as.dgRMatrix.spam( S) 
C <- as.dgCMatrix.spam( S)
as.spam.dgCMatrix(C) 
slotNames(U)
slotNames(R)
# For column oriented sparse formats a transpose does not the job,
# as the slot names change.

# as.spam(R) does not work.

## End(Not run)

## Not run: 
# for transformations between SparseM and spam:
as.matrix.csr.spam <- function(x,...) {
    newx <- new("matrix.csr")
    slot(newx,"ra",check=FALSE) <- x@entries
    slot(newx,"ja",check=FALSE) <- x@colindices
    slot(newx,"ia",check=FALSE) <- x@rowpointers
    slot(newx,"dimension",check=FALSE) <- x@dimension
    return(newx)
  }       
U <- as.matrix.csr.spam( S) 

## End(Not run)

## Not run: 
# a dataset contained in Matrix
data(KNex)
as.spam.dgCMatrix(KNex$mm) 

## End(Not run)

[Package spam version 2.5-1 Index]