germany.plot {spam}	R Documentation

Plot administrative districts of Germany

Description

Displaying data over the administrative districts of Germany

Usage

germany.plot(vect,  col=NULL, zlim=range(vect), legend=TRUE, 
             main=NULL, cex.axis=1, cex.main=1.5, add=FALSE, ... )

Arguments

vect
vector of length 544

col
color scheme to be used. By default uses colorRampPalette(brewer.pal(9,"Blues"))(100).

zlim
the minimum and maximum values for which colors should be plotted, defaulting to the range of data.

legend
Should the legend be added, see ‘Details’.

main
an overall title for the plot.

cex.axis
label size of legend.

cex.main
label size of overall plot title.

add
logical, if true adds to current plot.

...
additional arguments passed to polygon.

Details

The legend is only added, provided (a) function image.plot is available.

The perfect position of the legend is an art per se and depends on various par parameters. One possiblity for finer control is to not plot it and to manually call the function image.plot of fields.

Author(s)

Reinhard Furrer

References

See also http://de.wikipedia.org/wiki/Amtlicher_Gemeindeschl%C3%BCssel and http://de.wikipedia.org/wiki/Liste_der_Landkreise_in_Deutschland

See Also

Oral.

Examples

data( Oral)
germany.plot( Oral$Y/Oral$E)

# Plot the Bundeslaender:
germany.plot(germany.info$id%/%1000,col=rep(2:8,3), legend=FALSE)

[Package spam version 2.5-1 Index]