germany.data {spam}	R Documentation

Meta-data about administrative districts of Germany

Description

Data for the display of data over the administrative districts of Germany

Format

germany.info is a list with elements

n

544 (number of districts around 1990).

xrep,yrep

representative coordinates of the districts (vectors of length 544)

xlim,ylim

2-vectors defining the limits of the districts.

polyid

linking the polygons to the districts (599 vector).

id

linking the districts to Community Identification Number.

germany.poly defines the polygons. It is a 17965 by two matrix, each polygon separated by a row of NAs, each district by two rows.
germany defines the polygons in form of a list (backwards compatibility).

Details

The representative coordinates are calculated based on the mean value of the polygon coordinates. This creates sometimes strange values, e.g., district Leer.

Author(s)

Reinhard Furrer

References

The meta-data has been constructed based on (essentially) files from the package INLA, see demo(Bym).

See also http://de.wikipedia.org/wiki/Amtlicher_Gemeindeschl%C3%BCssel and http://de.wikipedia.org/wiki/Liste_der_Landkreise_in_Deutschland

See Also

germany.plot Oral.

Examples

# Plot the Bundeslaender:
germany.plot(germany.info$id%/%1000,col=rep(2:8,3), legend=FALSE)

[Package spam version 2.3-0.1 Index]