head {spam}	R Documentation

Return the First or Last Part of an Object

Description

Returns the upper left or lower right part of a spam object.

Usage

## S4 method for signature 'spam'
head(x, n = 6L, m = n, ...)
## S4 method for signature 'spam'
tail(x, n = 6L, m = n, addrownums = TRUE, ...)

Arguments

x
a spam object

n
a single integer. If positive, size for the resulting object: number of elements for a vector (including lists), rows for a matrix or data frame or lines for a function. If negative, all but the n last/first number of elements of x.

m
similar to n but for the number of columns.

addrownums
create row and column namves them from the selected elements.

...
arguments to be passed to or from other methods.

Details

For matrices, 2-dim tables and data frames, head() (tail()) returns the first (last) n rows and m columns when n > 0 or all but the last (first) n rows when n < 0 (with similar behavior for m).

tail() will add row and column names of the form "[n,]" and "[,n]" to the result, so that it looks similar to the last lines and columns of x when printed. Setting addrownums = FALSE suppresses this behaviour.

A method for spam.chol.NgPeyton objects is exported as well.

Value

An regular matrix.

Author(s)

Reinhard Furrer

Examples

head( precmat.RW2( 10))
tail( precmat.season(n=10, season=3), n=4, m=10)

[Package spam version 2.5-0 Index]