history {spam}	R Documentation

Brief Overview of the History

Description

Brief overview of spams history

Usage

  spam.history()

Details

Many years ago, I required sparse matrix algebra capacities in R for my research. The packages available at that time where not necessarily satisfying for my purpose. Hence, I started writing my own functions, somewhat inspired by the package SparseM.

Over the years, I decided to write an R package, called spam. In mean time, we have tree significant sparse matrix packages posted on CRAN. I believe that all three have their own niche of users.

I believe that spam is stable, useful but definitely not fully matured. I appreciate all comments and remarks.

There is no need to maintain here a complete list of changes. Older versions are available on CRAN (>0.11), each coming with a proper history file.

References

See also a more detailed description at http://www.math.uzh.ch/furrer/software/spam/history.shtml.

See Also

todo.

Examples

## Not run: 
## A complete history of changes
file.show(system.file("NEWS", package = "spam"))
file.show(system.file("0NEWS", package = "spam"))

## End(Not run)

[Package spam version 2.2-1 Index]