isSymmetric {spam}	R Documentation

Test if a spam matrix is Symmetric

Description

Efficient function to test if 'object' is symmetric or not.

Usage

# isSymmetric.spam(object, ...)
## S3 method for class 'spam'
isSymmetric(object, tol = 100 * .Machine$double.eps, ...)

Arguments

object
a spam matrix.

tol
numeric scalar >= 0. Smaller differences are not considered, see all.equal.spam.

...
further arguments passed to all.equal.spam.

Details

symmetry is assessed by comparing the sparsity structure of object and t(object) via the function all.equal.spam. If a difference is detected, the matrix is cleaned with cleanup and compared again.

Value

logical indicating if object is symmetric or not.

Author(s)

Reinhard Furrer

See Also

all.equal.spam, cleanup.

Examples

obj <- diag.spam(2)
isSymmetric(obj)

obj[1,2] <- .Machine$double.eps
isSymmetric(obj)
all.equal(obj, t(obj))

[Package spam version 2.2-1 Index]