map.landkreis {spam}	R Documentation

Administrative districts of Germany

Description

Displaying data over the administrative districts of Germany

Usage

map.landkreis(data, col=NULL, zlim=range(data), add=FALSE,
              legendpos=c( 0.88,0.9,0.05,0.4))

Arguments

data
vector of length 544

col
color scheme to be used. By default uses tim.colors if available or a generic gray scale.

zlim
the minimum and maximum values for which colors should be plotted, defaulting to the range of data.

add
logical, if true adds to current plot.

legendpos
if package fields is loaded, puts a legend at that position.

Details

The function germany.plot super-seeds map.landkreis (it is several factors faster).

The perfect position of the legend is an art per se and depends on various par parameters. See also the source code of the function image.plot of fields.

Author(s)

Reinhard Furrer

References

The code of map.landkreis is very similar to germany.map from the package INLA.

See Also

germany.plot super-seeding map.landkreis.

Examples

## Not run: 
data( Oral)
par( mfcol=c(1,2))
germany.plot( log( Oral$Y), legend=TRUE)
map.landkreis( log( Oral$Y))

## End(Not run)

[Package spam version 2.3-0.1 Index]