Math {spam}	R Documentation

Mathematical functions

Description

Applies the Math group functions to spam objects

Usage

# ceiling(x)
# floor(x)
# exp(x, base = exp(1))
# log(x, base = exp(1))
# sqrt(x)

# abs(x)
# cumprod(x)
# cumsum(x)

# cos(x)
# sin(x)
# tan(x)

# acosh(x)

Arguments

x
spam object.

base
positive number. The base with respect to which logarithms are computed. Defaults to e=exp(1).

Details

It is important to note that the zero entries do not enter the evaluation when structurebased=FALSE. The operations are performed on the stored non-zero elements. This may lead to differences if compared with the same operation on a full matrix.

Value

If the option spam.structurebased=TRUE, all functions operate on the vector x@entries and return the result thereof.
Conversely, if structurebased=FALSE, the result is identical to one with as.matrix(x) input and an as.spam purger.

Author(s)

Reinhard Furrer

See Also

Summary.spam, Ops.spam and Math2.spam

Examples

getGroupMembers("Math")

mat <- matrix(c( 1,2,0,3,0,0,0,4,5),3)
smat <- as.spam( mat)
cos( mat)
cos( smat)

options(spam.structurebased=FALSE)
cos( smat)

sqrt( smat)

[Package spam version 2.3-0.1 Index]