spam methods {spam}	R Documentation

Methods for sparse matrices

Description

Methods without any additional parameters for sparse matrices.

Details

If a method for spam objects takes the same arguments, produces the intuitive output. We do not provide additional help pages.

However, such methods are usually linked to a xzy.spam function, that could also be called directly.

Author(s)

Reinhard Furrer

See Also

Corresponding base help functions.

[Package spam version 2.3-0.1 Index]