random {spam}	R Documentation

Create Random spam Matrices

Description

Creates random spam matrix given the dimension and other parameters.

Usage

spam_random(nrow = 1L, ncol = nrow, density = 0.5, distribution = NULL, digits = NULL,
            sym = FALSE, spd = FALSE, verbose = FALSE, ...)

Arguments

nrow
integer value for the number of rows for the spam matrix to create.

ncol
integer value for the number of columns. The default value is the same as nrow.

density
A numeric value between 0 and 1 specifying the approximate density of matrix. If equal to zero the spam matrix contains only zeros and if equal to 1 the spam matrix is full.

distribution
a random number generating distribution function to sample the entries of the spam matrix. The function must have an argument with the name n, possible candidates are rnorm, rexp, rpois, rweibull, etc..

...
possible additional arguments for the distribution function if specified with distribution.

digits
an integer value for the number of digits the entries should be rounded.

sym
logical value to specify symmetry of the spam matrix.

spd
logical value to specify positive definitness of the spam matrix, via diagonal dominace criteria. Note, if spd TRUE, then sym is overwritten to TRUE in any case.

verbose
logical value to specify verbose statments of the function.

Details

To create a random spam64 matrix, set options(spam.force64 = TRUE).

Value

A random matrix in spam format.

Author(s)

Florian Gerber, Roman Flury, Reinhard Furrer

See Also

spam and display.spam

Examples

set.seed(42)
rspam <- spam_random(500, digits = 2, distribution = rnorm, sd = 2, mean = 10, density = .01)
display.spam(rspam, cex = 2)

[Package spam version 2.5-1 Index]