s3only {spam}	R Documentation

Wappers for Sparse Matrices

Description

These functions are convenient wrappers for spam objects to classical matrix operations.

Usage

var.spam(x, ...)

## S3 method for class 'spam'
var(x, ...)

Arguments

x
matrix of class spam.

...
further arguments passed to or from other methods.

Details

There is probably no point in fully defining methods here. Typically, these functions do not exploit sparsity structures. Hence, for very large matrices, warnings may be posted.

Value

Depends on function...

Author(s)

Reinhard Furrer

See Also

Option "inefficiencywarning" in spam.options.

[Package spam version 2.3-0.1 Index]