spam.chol.NgPeyton-class {spam}	R Documentation

Class "spam.chol.NgPeyton"

Description

Result of a Cholesky decomposition with the NgPeyton method

Objects from the Class

Objects are created by calls of the form chol(x,method="NgPeyton", ...) and should not be created directly with a new("spam.chol.NgPeyton", ...) call.
At present, no proper print method is defined. However, the factor can be transformed into a spam object.

Methods

as.matrix

signature(x = "spam.chol.NgPeyton"): Transform the factor into a regular matrix.

as.spam

signature(x = "spam.chol.NgPeyton"): Transform the factor into a spam object.

backsolve

signature(r = "spam.chol.NgPeyton"): solving a triangular system, see solve.

forwardsolve

signature(l = "spam.chol.NgPeyton"): solving a triangular system, see solve.

c

signature(x = "spam.chol.NgPeyton"): Coerce the factor into a vector.

determinant

signature(x = "spam.chol.NgPeyton"): Calculates the determinant from the factor, see also det.

diag

signature(x = "spam.chol.NgPeyton"): Extracts the diagonal entries.

dim

signature(x = "spam.chol.NgPeyton"): Retrieve the dimension. Note that "dim<-" is not implemented.

display

signature(x = "spam.chol.NgPeyton"): Transformation to a spam object and display, see also display.

image

signature(x = "spam.chol.NgPeyton"): Transformation to a spam object and display, see also image.

length

signature(x = "spam.chol.NgPeyton"): Retrieve the dimension. Note that "length<-" is not implemented.

ordering

signature(x = "spam.chol.NgPeyton"): Retrieves the ordering, in ordering.

print

signature(x = "spam.chol.NgPeyton"): Short description.

show

signature(object = "spam.chol.NgPeyton"): Short description.

summary

signature(object = "spam.chol.NgPeyton"): Description of the factor, returns (as a list) nnzR, nnzcolindices, the density of the factor density, and fill-in ratio fillin. For the use of the first two, see ‘Examples’ in chol.

t

signature(x = "spam.chol.NgPeyton"): Transformation to a spam object and transposition.

chol

signature(x = "spam.chol.NgPeyton"): Returns x unchanged.

Author(s)

Reinhard Furrer

References

Ng, E. G. and B. W. Peyton (1993), "Block sparse Cholesky algorithms on advanced uniprocessor computers", SIAM J. Sci. Comput., 14, pp. 1034-1056.

See Also

print.spam ordering and chol

Examples

x <- spam( c(4,3,0,3,5,1,0,1,4),3)
cf <- chol( x)
cf
as.spam( cf)

# Modify at own risk...
slotNames(cf)

[Package spam version 2.5-0 Index]