spam internal {spam}	R Documentation

Spam internal and auxiliary functions

Description

The functions listed below are auxiliary functions but are exported by the NAMESPACE. The user should not require to call these directly.

Details

The functions are listed here for a better understanding of the code (to fulfill the tutorial style paradigm).

validspamobject(object) A few sanity checks if object is a proper spam object.

dcheck(x), icheck(x) testing and forcing of x to doubles and integers. These functions are used when calling Fortran routines.

[Package spam version 2.3-0.1 Index]