Summary {spam}	R Documentation

Rounding of Numbers

Description

Applies the Math2 group functions to spam objects

Usage

# max(x,..., na.rm = FALSE)

Arguments

x
spam object.

na.rm
a logical indicating whether missing values should be removed.

Details

The na.rm argument is only meaninful if NAOK=TRUE.

Value

If structurebased=TRUE, all functions operate on the vector x@entries and return the result thereof.
Conversely, if structurebased=FALSE, the result is identical to one with as.matrix(x) input.

Author(s)

Reinhard Furrer

See Also

Math.spam and Math2.

Examples

getGroupMembers("Summary")

smat <- diag.spam( runif(15))
range(smat)
options(spam.structurebased=FALSE)
range(smat)

## Not run: 
max( log(spam(c(1,-1))), na.rm=TRUE)

## End(Not run)
# allow 'NA's first:
# TODO
# options(spam.NAOK=TRUE)
# max( log(spam(c(1,-1))), na.rm=TRUE)

[Package spam version 2.5-1 Index]