todo {spam}	R Documentation

Small "ToDo" list

Description

List of what needs to be done within spam

Usage

  todo()

Details

This is a non exhaustive list of where we need to work on spam (of course the list is in random order):

- extend demo(s)

- write vignette

- complete help files

- extend basic matrix operatation, comparisons, etc:
unique, duplicated, ...

- improve subsetting via row extraction, incorporate matrix permutation

- implement other Cholesky routines (one eye glances to the LDL library).

- what about an LU/SVD decomposition?

Any other items are welcome (reinhard.furrer@math.uzh.ch).

[Package spam version 2.2-1 Index]