toeplitz {spam}	R Documentation

Create Toeplitz Matrices

Description

Creates symmetric and asymmetric Toeplitz matrices.

Usage

toeplitz.spam(x, y = NULL, eps = getOption("spam.eps"))

Arguments

x
the first row to form the Toeplitz matrix.

y
for asymmetric Toeplitz matrices, this contains the first column.

eps
A tolerance parameter: elements of x such that abs(x) <= eps set to zero. Defaults to eps = getOption("spam.eps").

Details

The vector y has to be of the same length as x and its first element is discarded.

Value

The Toeplitz matrix in spam format.

Author(s)

Reinhard Furrer

See Also

toeplitz, circulant.spam

Examples

toeplitz.spam(c(1,.25,0,0,0))

[Package spam version 2.5-0 Index]