version {spam}	R Documentation

Spam Version Information

Description

spam.version is a variable (list) holding detailed information about the version of spam loaded.

spam.Version() provides detailed information about the version of spam running.

Usage

spam.version

Value

spam.version is a list with character-string components

status
the status of the version (e.g., "beta")

major
the major version number

minor
the minor version number

year
the year the version was released

month
the month the version was released

day
the day the version was released

version.string
a character string concatenating the info above, useful for plotting, etc.

spam.version is a list of class "simple.list" which has a print method.

Author(s)

Reinhard Furrer

See Also

See the R counterparts R.version.

Examples

spam.version$version.string

[Package spam version 2.2-1 Index]